requirements.txt
Django>=4.2,<5.0
psycopg2-binary
python-dotenv
django-crispy-forms
crispy-tailwind
djangorestframework
reportlab
openai
gunicorn
